requirements.txt
google-cloud-storage
google-cloud-bigquery
functions-framework
sendgrid
python_http_client